Separate CSV validation tests
The package verifier tests mixed archive layout, rights metadata, and transfer
metadata in one large module. The expanded rights rules made that module
increasingly difficult to navigate.

Rights and metadata scenarios now live in dedicated test modules, while the
original module retains package-level checks. This is a test-only
reorganization that preserves the existing coverage.
